+++ addendum/management/commands/refresh_snippet_cache.py
@@ -4,20 +4,17 @@ Management command to update all snippet cached values.
 
 from django.core.management.base import BaseCommand
 
-from ...models import SnippetTranslation, Snippet
+from ...models import set_cached_snippet, Snippet
 
 
 class Command(BaseCommand):
-    help = "Updates the cache values for all snippets including translations"
+    help = "Updates the cache values for all snippets"
 
     def handle(self, *args, **kwargs):
-        snippet_count, translation_count = 0, 0
+        # Already iterating, skip the extra count query
+        count = 0
         for snippet in Snippet.objects.all():
-            snippet.set_cache()
-            snippet_count += 1
-        for snippet in SnippetTranslation.objects.all():
-            snippet.set_cache()
-            translation_count += 1
+            set_cached_snippet(snippet.key)
+            count += 1
         self.stdout.write(
-                "Refreshed the cache for {0} snippets and {1} translations".format(
-                    snippet_count, translation_count))
+                "Refreshed the cache for {0} snippets.".format(count))